+++ TwitchChatOverlay/Infrastructure/RegionNames.cs
@@ -1,0 +1,9 @@
+namespace TwitchChatOverlay.Infrastructure
+{
+    public static class RegionNames
+    {
+        public const string CommonSettingsRegion = "CommonSettingsRegion";
+        public const string TwitchSettingsRegion = "TwitchSettingsRegion";
+        public const string YouTubeSettingsRegion = "YouTubeSettingsRegion";
+    }
+}