+++ TwitchChatOverlay/Views/Tabs/CommonSettingsTabView.xaml.cs
@@ -1,0 +1,12 @@
+using System.Windows.Controls;
+
+namespace TwitchChatOverlay.Views.Tabs
+{
+    public partial class CommonSettingsTabView : UserControl
+    {
+        public CommonSettingsTabView()
+        {
+            InitializeComponent();
+        }
+    }
+}