+++ TwitchChatOverlay/Views/Tabs/YouTubeSettingsTabView.xaml.cs
@@ -1,0 +1,12 @@
+using System.Windows.Controls;
+
+namespace TwitchChatOverlay.Views.Tabs
+{
+    public partial class YouTubeSettingsTabView : UserControl
+    {
+        public YouTubeSettingsTabView()
+        {
+            InitializeComponent();
+        }
+    }
+}